Move some fixtures into conftest.py for broader use
